+++ src/main/java/com/braintreegateway/PaymentMethodNonceDetails.java
@@ -5,10 +5,12 @@ import com.braintreegateway.util.NodeWrapper;
 public class PaymentMethodNonceDetails {
     private String cardType;
     private String lastTwo;
+    private String lastFour;
 
     public PaymentMethodNonceDetails(NodeWrapper node) {
         cardType = node.findString("card-type");
         lastTwo = node.findString("last-two");
+        lastFour = node.findString("last-four");
     }
 
     public String getCardType() {
@@ -19,4 +21,7 @@ public class PaymentMethodNonceDetails {
         return lastTwo;
     }
 
+    public String getLastFour() {
+        return lastFour;
+    }
 }